Replace the previous implementation with the new files.
